We love you sir

+++ Project/src/school/management/system/Student.java
@@ -7,7 +7,7 @@ public class Student {
 	private int feesPaid;
 	private int feesTotal;
 	
-	public Student(int id,String name,double grade) {
+	public Student(int id,String name,int grade) {
 		this.feesPaid=0;
 		this.feesTotal=30000;
 		this.id=id;
@@ -16,7 +16,7 @@ public class Student {
 		
 		
 	}
-   public void setGrade(double grade)
+   public void setGrade(int grade)
    {
 	   this.grade=grade;
    }
